- Show exclude pokemon subs with individual values list set - Fix gym image url DTS -
