+++ examples/Alerts/default.json
@@ -18,9 +18,9 @@
         "imageUrl": "<tilemaps_url>"
     },
     "gyms": {
-        "avatarUrl": "<gym_img_url>",
+        "avatarUrl": "<gym_url>",
         "content": "<#team_changed>Gym changed from <old_gym_team_emoji> <old_gym_team> to <gym_team_emoji> <gym_team><br></team_changed><#in_battle>Gym is under attack!<br></in_battle>**Slots Available:** <slots_available><br><#is_ex><ex_gym_emoji> Gym!</is_ex>**[[Google Maps](<gmaps_url>)] [[Apple Maps](<applemaps_url>)] [[Waze Maps](<wazemaps_url>)]**",
-        "iconUrl": "<gym_img_url>",
+        "iconUrl": "<gym_url>",
         "title": "<geofence>: <gym_name>",
         "url": "<gmaps_url>",
         "username": "<gym_name>",
