+++ src/Extensions/IntegerExtensions.cs
@@ -4,11 +4,12 @@
 
     public static class IntegerExtensions
     {
-
+        /*
         public static char NumberToAlphabet(this int num)
         {
             return Convert.ToChar(num + 64);
         }
+        */
 
         public static DateTime FromUnix(this long unixSeconds)
         {
