+++ src/Extensions/StringExtensions.cs
@@ -29,7 +29,6 @@
                     }
 
                     msg = msg.Replace("{" + i + "}", args[i].ToString());
-                    msg = msg.Replace("\n", "\n");
                 }
                 return msg;
             }
@@ -52,13 +51,6 @@
             {
                 yield return s.Substring(i, Math.Min(partLength, s.Length));
             }
-        }
-
-        public static string ToUppercaseFirstLetter(this string text)
-        {
-            var firstLetter = text[0].ToString().ToUpper();
-            var remainingLetters = text.Substring(1, text.Length - 1);
-            return firstLetter + remainingLetters;
         }
     }
 }
