+++ src/Net/Models/TeamRocketInvasion.cs
@@ -2,8 +2,11 @@
 {
     using System;
     using System.Collections.Generic;
+    using System.Linq;
 
     using Newtonsoft.Json;
+
+    using WhMgr.Data;
 
     public class TeamRocketInvasion
     {
@@ -25,6 +28,35 @@
         public TeamRocketInvasion()
         {
             Encounters = new TeamRocketEncounters();
+        }
+
+        public string GetPossibleInvasionEncounters()
+        {
+            var toInt = new Func<string, int>(x =>
+            {
+                var val = x.Split('_')[0];
+                if (!int.TryParse(val, out var result))
+                {
+                    Console.Error.WriteLine($"Failed to parse {val} as integer");
+                }
+                return result;
+            });
+            var first = string.Join(", ", Encounters.First.Select(x => MasterFile.GetPokemon(toInt(x), 0)?.Name));
+            var second = string.Join(", ", Encounters.Second.Select(x => MasterFile.GetPokemon(toInt(x), 0)?.Name));
+            //var third = string.Join(", ", invasion.Encounters.Third.Select(x => Database.Instance.Pokemon[x].Name));
+            var msg = string.Empty;
+            if (SecondReward)
+            {
+                //85%/15% Rate
+                msg += $"85% - {first}\r\n";
+                msg += $"15% - {second}\r\n";
+            }
+            else
+            {
+                //100% Rate
+                msg += $"100% - {first}\r\n";
+            }
+            return msg;
         }
     }
